Function yatp::pool::build_spawn

pub fn build_spawn<T>(
    queue_type: impl Into<QueueType>, 
    config: SchedConfig
) -> (Remote<T>, Vec<Local<T>>) where
    T: TaskCell + Send, 

Building remotes and locals from the given queue and configuration.

This is only for tests purpose so that a thread pool doesn't have to be spawned to test a Runner.